+++ components/PagedTable.js
@@ -53,10 +53,6 @@ function PagedTable({ cols, rows, n, rowHandler = () => {}, moreUrl, lazy }) {
     return rows
   }
 
-  // useEffect(() => {
-  //   console.log(viewSet)
-  // }, [viewSet])
-
   return (
     <div className="app-table">
       <table>